@@ -36,37 +36,32 @@
     "require": {
         "php": ">=5.6.0",
         "acclimate/container": "^1.0",
-        "franzl/whoops-middleware": "^0.2.0",
-        "illuminate/database": "^5.1",
-        "league/container": "^2.0.3",
-        "league/factory-muffin": "^2.1",
+        "franzl/whoops-middleware": "^0.5.0",
+        "illuminate/database": "^5.4",
+        "league/container": "^2.4",
+        "league/factory-muffin": "^3.0",
         "league/flysystem": "^1.0",
-        "league/route": "2.0.0-RC1",
-        "league/tactician": "^0.6.1",
-        "madewithlove/definitions": "1.0.x@dev",
-        "maximebf/debugbar": "^1.10",
+        "league/route": "^2.0",
+        "league/tactician": "^1.0",
+        "madewithlove/service-providers": "^0.3.0",
+        "maximebf/debugbar": "^1.13",
         "mnapoli/assembly": "^0.2.0",
-        "monolog/monolog": "^1.17",
-        "oscarotero/psr7-middlewares": "^3.8",
-        "psy/psysh": "^0.6.1",
-        "relay/relay": "^1.0",
-        "robmorgan/phinx": "^0.5.0",
-        "symfony/console": "^2.8",
-        "twig/twig": "^1.23",
-        "vlucas/phpdotenv": "^2.1",
-        "willdurand/negotiation": "^2.0",
-        "zendframework/zend-diactoros": "^1.2"
+        "monolog/monolog": "^1.22",
+        "oscarotero/psr7-middlewares": "^3.19",
+        "psy/psysh": "^0.8.2",
+        "relay/relay": "^1.1",
+        "robmorgan/phinx": "^0.8.0",
+        "symfony/console": "^3.2",
+        "twig/twig": "^2.2",
+        "vlucas/phpdotenv": "^2.4",
+        "willdurand/negotiation": "^2.2",
+        "zendframework/zend-diactoros": "^1.3"
     },
     "require-dev": {
-        "fabpot/php-cs-fixer": "2.0.x@dev",
-        "illuminate/container": "^5.1",
-        "mockery/mockery": "^0.9.4",
-        "phpunit/phpunit": "^5.0.9"
-    },
-    "extra": {
-        "branch-alias": {
-            "dev-master": "1.0-dev"
-        }
+        "madewithlove/php-cs-fixer-config": "^1.2",
+        "illuminate/container": "^5.4",
+        "mockery/mockery": "^0.9.9",
+        "phpunit/phpunit": "^6.0"
     },
     "scripts": {
         "test": "phpunit",
@@ -78,5 +73,14 @@
             "cd _book && git init && git checkout -b gh-pages && touch .nojekyll && git add --all && git commit -am 'Rebuild documentation' && git push git@github.com:madewithlove/glue.git gh-pages --force"
         ]
     },
-    "minimum-stability": "stable"
+    "extra": {
+        "branch-alias": {
+            "dev-master": "1.0-dev"
+        }
+    },
+    "minimum-stability": "dev",
+    "prefer-stable": true,
+    "config": {
+        "sort-packages": true
+    }
 }
